Bug fixes discovered during session and also fix restore.

+++ Actions/DeleteMoveUnitAction.cs
@@ -27,14 +27,17 @@ public class DeleteMoveUnitAction : PlayerAction
         {
             MapHex mapHex = gameState.Map.Hexes[Unit.Y, Unit.X];
             Unit existingUnit = mapHex.getUnit();
-            existingUnit.DeleteMoveUnitActions();
-            existingUnit.IsLoading = false;
-            existingUnit.IsUnloading = false;
-            existingUnit.RoundsToPause = 0;
-            existingUnit.UnitToPursueX = -1;
-            existingUnit.UnitToPursueY = -1;
-            existingUnit.UnitIdToPursue = null;
-            server.sendGameStateAndMapHex(Unit.X, Unit.Y);
+            if (existingUnit != null)
+            {
+                existingUnit.DeleteMoveUnitActions();
+                existingUnit.IsLoading = false;
+                existingUnit.IsUnloading = false;
+                existingUnit.RoundsToPause = 0;
+                existingUnit.UnitToPursueX = -1;
+                existingUnit.UnitToPursueY = -1;
+                existingUnit.UnitIdToPursue = null;
+                server.sendGameStateAndMapHex(Unit.X, Unit.Y);
+            }
         }
     }
 }
